+++ SmartHall/SmartHall.Contracts/Halls/SearchFreeHall/SearchFreeHallRequest.cs
@@ -1,4 +1,4 @@
 ﻿namespace SmartHall.Contracts.Halls.GetFreeHall
 {
-	public record SearchFreeHallRequest(DateTime DateTime, double Hours, int Capacity);
+	public record SearchFreeHallRequest(DateTime DateTime, int Hours, int Capacity);
 }
